Re-create a client in the RoundTripper
The RoundTripper stores a map of clients that use the hostname as key.
If a client in this map has some problem, like a timeout (Destroy
session: NO_ERROR: No recent network activity), then the RoundTripper
needs to be recreated otherwise is impossible to send more requests to
the same hostname.

With this change, I'm trying to detect if the underline session is done
and create a new client for the same hostname.